+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.4.2] - 2020-11-08
+
+### Changed
+
+- Texts on the English home page.
+- Skills and knowledge to be presented in a clearer form in the curriculum vitae.
+
 ## [0.4.1] - 2020-11-08
 
 ### Removed
@@ -83,7 +90,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 - Existing contents of the site from the old source code repository at [anttikivi/shared-sonnet](https://github.com/anttikivi/shared-sonnet).
 
-[unreleased]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.4.1...HEAD
+[unreleased]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.4.2...HEAD
+[0.4.2]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/anttikivi/anttikivi.github.io/compare/v0.3.0...v0.3.1